check that amounts are equal in the equality method of Dollar.
At this point in the book, the 'other' argument is cast to a Dollar before
checking amount. Since I'm not terribly familiar with Java, I'm not sure if
the cast would prevent the possibility of a runtime error if passed an object
that wouldn't be convertable to Dollar. "casting" doesn't make much sense in
Python, so I chose to just try to compare against 'other's amount.

A simple if not "hasattr(other, 'amount') return NotImplemented" would be an
appropriate check, but isn't implemented here since it would be breaking the
principles being put forth in the book, and because it may turn up later.