Add a failing test showing that adding two Money's should produce a Total.
We're using Total instead of Sum here to avoid minor confusion with python's
"sum" builtin.

The code here in the book is relying on interfaces and casting. We don't really
have those at the language level in Python, so we attempt to approximate the idea
by passing "result" into the constructor of Total. This is awkward.